Fix: agent: prevent gethosts action and timeout validation from hanging on list and dump commands if any of the devices is silently blocked
If any of the configured SBD devices is silently blocked without any
explicit I/O error from kernel, fencing will get stuck and time out,
even if the majority of the devices are still available.

On fencing, list and dump commands are called first. Under this
situation, the commands will print output but get stuck on exit_aio() on
exit, and become D state.

With this commit, sbd fence agent asynchronously calls the commands
individually for the devices and wait for any successful return and
collect the output, so that it prevents execution of sbd fence agent
from hanging.